backend/requirements.txt
agno
flask
flask-cors
numpy
python-dotenv
reportlab
google-generativeai
